Implemented location attack code :rocket:
With each request to `/login` endpoint, we check the `locationHistory`
collection in firestore for that user. If that user has not logged in
from that locatin in the past, we generate a token and save the new
location.

The end user must then go to the `/verifyLocation` endpoint with the
token in the request query. If it matches, we add the new location to
the location history and the user can then access their credentials.